klocwork#255 Fix sprintf(%d, u_int32_t)

+++ servers/slapd/back-bdb/config.c
@@ -387,8 +387,8 @@ bdb_cf_gen( ConfigArgs *c )
 			if ( bdb->bi_txn_cp ) {
 				char buf[64];
 				struct berval bv;
-				bv.bv_len = snprintf( buf, sizeof(buf), "%d %d", bdb->bi_txn_cp_kbyte,
-					bdb->bi_txn_cp_min );
+				bv.bv_len = snprintf( buf, sizeof(buf), "%ld %ld",
+					(long) bdb->bi_txn_cp_kbyte, (long) bdb->bi_txn_cp_min );
 				if ( bv.bv_len > 0 && bv.bv_len < sizeof(buf) ) {
 					bv.bv_val = buf;
 					value_add_one( &c->rvalue_vals, &bv );
